Move MedianAnalogInput into experimental.knobs
